Project Description, Project Scope, License, A couple of data structures
I wrote some data structures out in order to properly control the
traffic lights. The two structures I wrote are for the traffic lights
and the traffic cameras. These structures aren’t complete. What are
complete are the license terms, the project scope, and project
description. This will be used to guide the rest of the project.